Refactor location handlers to submodules
