+++ server/api/locations/location/handlers/changeName.js
@@ -1,0 +1,36 @@
+const locationDal = require('../dal');
+const status = require('http-status-codes');
+
+module.exports = (req, res, next) => {
+  // Update location name
+
+  let newLocName = req.body.newName;
+
+  if (typeof newLocName !== 'string') {
+    return res.status(status.BAD_REQUEST).send('Invalid name');
+  }
+
+  // Remove excess whitespaces and prevent only-whitespace names.
+  newLocName = newLocName.trim();
+
+  const minNameLen = 2;
+  const maxNameLen = 120;
+  if (newLocName.length < minNameLen || newLocName.length > maxNameLen) {
+    return res.status(status.BAD_REQUEST).send('Too short or too long name');
+  }
+
+  const params = {
+    locationId: req.location._id,
+    locationName: req.location.name,
+    newName: req.body.newName,
+    username: req.user.name,
+  };
+
+  locationDal.changeName(params, (err) => {
+    if (err) {
+      return next(err);
+    }
+
+    return res.sendStatus(status.OK);
+  });
+};
