+++ server/api/locations/location/handlers/index.js
@@ -1,0 +1,8 @@
+exports.changeGeom = require('./changeGeom');
+exports.changeName = require('./changeName');
+exports.changeStatus = require('./changeStatus');
+exports.changeTags = require('./changeTags');
+exports.changeThumbnail = require('./changeThumbnail');
+exports.changeType = require('./changeType');
+exports.getOne = require('./getOne');
+exports.removeOne = require('./removeOne');
